fix: resolve compiler warnings for clean -Wall -Wextra -Wshadow build
Add configurable compiler warning infrastructure (cmake/StrictCompilerWarnings.cmake)
with 3 warning levels and optional -Werror support. Fix all warnings across the
codebase including -Wshadow, -Wunused-parameter, -Wnon-virtual-dtor, -Wformat,
-Wimplicit-fallthrough, and -Woverloaded-virtual warnings in 160+ source files.

The library and tests now compile cleanly at warning level 2 (-Wall -Wextra
-Wpedantic -Wshadow -Wnon-virtual-dtor -Wunused -Woverloaded-virtual -Wformat=2
-Wimplicit-fallthrough) with -Werror enabled.

+++ faiss/utils/partitioning.cpp
@@ -50,8 +50,8 @@ typename C::T sample_threshold_median3(
     T val3[3];
     int vi = 0;
 
-    for (size_t i = 0; i < n; i++) {
-        T v = vals[(i * big_prime) % n];
+    for (size_t i = 0; i < static_cast<size_t>(n); i++) {
+        T v = vals[(i * big_prime) % static_cast<size_t>(n)];
         // thresh_inf < v < thresh_sup (for CMax)
         if (C::cmp(v, thresh_inf) && C::cmp(thresh_sup, v)) {
             val3[vi++] = v;
@@ -1271,14 +1271,14 @@ void simd_histogram_16(
         int* hist) {
     memset(hist, 0, sizeof(*hist) * 16);
     if (shift < 0) {
-        for (size_t i = 0; i < n; i++) {
+        for (size_t i = 0; i < static_cast<size_t>(n); i++) {
             hist[data[i]]++;
         }
     } else {
         int vmax0 = std::min((16 << shift) + min, 65536);
         uint16_t vmax = uint16_t(vmax0 - 1 - min);
 
-        for (size_t i = 0; i < n; i++) {
+        for (size_t i = 0; i < static_cast<size_t>(n); i++) {
             uint16_t v = data[i];
             v -= min;
             if (!(v <= vmax))
@@ -1304,11 +1304,11 @@ void simd_histogram_8(
         int* hist) {
     memset(hist, 0, sizeof(*hist) * 8);
     if (shift < 0) {
-        for (size_t i = 0; i < n; i++) {
+        for (size_t i = 0; i < static_cast<size_t>(n); i++) {
             hist[data[i]]++;
         }
     } else {
-        for (size_t i = 0; i < n; i++) {
+        for (size_t i = 0; i < static_cast<size_t>(n); i++) {
             if (data[i] < min)
                 continue;
             uint16_t v = data[i] - min;
